% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.fdr.R
\name{combined.fdr}
\alias{combined.fdr}
\title{Calculating combined false positive rate based on Monte Carlo sampling}
\usage{
combined.fdr(sens, specs, prev, size = 1e+06)
}
\arguments{
\item{sens}{A vector of K sensitivity values for K different testing methods.}

\item{specs}{A vector of K specificity values for K different testing methods, with the order of the methods same as in \code{sens}.}

\item{prev}{Prevalence of true positives.}

\item{size}{A large integer for Monte Carlo sampling. Default: 1 million.}
}
\value{
FDR estimates for each category of testing results combination.
}
\description{
The function returns the estimated false discovery rate for each category of testing results, given the operating characteristics of different testing methods.
}
\note{
The function is based on Monte Carlo sampling estimation. Make sure a sufficiently large \code{size} is given.
}
\examples{
\dontrun{

combined.fdr(c(.5,.6,.7), c(.5,.4,.3), .5)
#        000        001        010        011        100        101        110        111
# 0.50135036 0.50012141 0.49931784 0.49982606 0.49945863 0.49953203 0.50014472 0.50040443
}

}
\references{
Yang Z, Xu W, Zhai R, Li T, Ning Z, Pawitan Y, Shen X (2020). Triangulation of analysis strategies links complex traits to specific tissues and cell types. 
\emph{Submitted}.
}
\seealso{
\code{triangulate}
}
\author{
Zhijian Yang, Xia Shen
}
