% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood of sensitivity, specificity, and prevalence}
\usage{
loglik(param, ntest, ny, tol = 1e-04)
}
\arguments{
\item{param}{A vector of 2K + 1 parameter values, corresponding to K sensitivity values, K specificity values, and prevalence of true positives, respectively.}

\item{ntest}{An integer of the number of tests performed by each method, i.e., sample size for evaluating the likelihood.}

\item{ny}{A summary counts vector for the contingencies of each combination of test results, must be named according to the 0-1 pattern, 
e.g., 010 = results from 3 tests are negative, positive, and negative, respectively. \code{names(ny)} maybe incomplete and has arbitrary order.}

\item{tol}{A value of tolerance, default 1e-4. If any \code{param} value or marginal probability is too close to 0 or 1, distance less than \code{tol}, 
a log-likelihood value of \code{exp(30)} will be returned.}
}
\value{
A single value of minus log-likelihood.
}
\description{
The function returns the minus log-likelihood value given sensitivity and specificity of at least three testing methods and prevalence of true positives as parameters, 
and contingency table of testing results of the methods as data.
}
\note{
This is an internal function for likelihood calculation.
}
\examples{
\dontrun{
sens <- c(.5,.6,.7)
spec <- c(.5,.4,.3)
p <- .5
param <- c(sens, spec, p)
ny <- c(123, 456, 789)
names(ny) <- c('010', '110', '111')
ny
# 010 110 111
# 123 456 789

loglik(param, ntest = 3, ny)
# [1] 2844.676
}

}
\references{
Yang Z, Xu W, Zhai R, Li T, Ning Z, Pawitan Y, Shen X (2020). Triangulation of analysis strategies links complex traits to specific tissues and cell types. 
\emph{Submitted}.
}
\seealso{
\code{triangulate}
}
\author{
Xia Shen, Yudi Pawitan
}
