% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixprob.R
\name{mixprob}
\alias{mixprob}
\title{Marginal probability as a function of sensitivity, specificity, and prevalence}
\usage{
mixprob(sens, spec, p)
}
\arguments{
\item{sens}{A vector of K sensitivity values for K different testing methods.}

\item{spec}{A vector of K specificity values for K different testing methods, with the order of the methods same as in \code{sens}.}

\item{p}{Prevalence of true positives.}
}
\value{
A named vector of probabilities for the categories of testing results. The naming is from '00...0' to '11...1', with length K for the K methods, 
0-1 pattern corresponds to the test results, e.g., 010 = results from 3 tests are negative, positive, and negative, respectively. 
The order is the same as in input \code{sens} and \code{spec}.
}
\description{
The function returns the marginal probability values based on sensitivity and specificity of at least three testing methods and prevalence of true positives.
}
\note{
This is an internal function for likelihood calculation.
}
\examples{
\dontrun{
mixprob(sens = c(.5,.6,.7), spec = c(.5,.4,.3), p = .5)

#  000  100  010  110  001  101  011  111
# 0.06 0.06 0.09 0.09 0.14 0.14 0.21 0.21
}

}
\references{
Yang Z, Xu W, Zhai R, Li T, Ning Z, Pawitan Y, Shen X (2020). Triangulation of analysis strategies links complex traits to specific tissues and cell types. 
\emph{Submitted}.
}
\seealso{
\code{loglik}
}
\author{
Xia Shen, Yudi Pawitan
}
