% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\docType{data}
\name{shape}
\alias{shape}
\title{Example from Distmesh (MATLAB)}
\format{
\code{shape} is a dataframe with vertices of the boundary of a polygon region.
\describe{
  \item{v1}{The latitudes of the external boundary.}
  \item{v2}{The longtitudes of the external boundary.}
}
}
\usage{
data('shape')
}
\description{
A dataframe of coordinates of the external boundary of a polygon region developed.
}
\references{
This example is from a triangulation software called distmesh.
}
\keyword{datasets}
