% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate.R
\name{triangulate}
\alias{triangulate}
\title{Maximum likelihood estimation of sensitivity, specificity, and prevalence}
\usage{
triangulate(
  counts,
  ntest,
  method = "Nelder-Mead",
  B = 0,
  start = c(rep(0.5, ntest * 2), 0.5)
)
}
\arguments{
\item{counts}{A summary counts vector for the contingencies of each combination of test results, must be named according to the 0-1 pattern, 
e.g., 010 = results from 3 tests are negative, positive, and negative, respectively. \code{names(ny)} maybe incomplete and has arbitrary order.}

\item{ntest}{An integer of the number of tests performed by each method, i.e., sample size for evaluating the likelihood.}

\item{method}{The optimization method to be used, can be one of "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", or "Brent". See Note. 
e.g., 010 = results from 3 tests are negative, positive, and negative, respectively. \code{names(ny)} maybe incomplete and has arbitrary order.}

\item{B}{An integer for the number of bootstrap samples, default 0 means no bootstrap estimation for standard errors.}

\item{start}{A vector of 2K + 1 parameter starting values, corresponding to K sensitivity values, K specificity values, and prevalence of true positives, respectively. Default: 0.5 for all parameters.}
}
\value{
A list of parameter estimates. If B > 0, the list also contains the standard errors obtained via bootstrap.
}
\description{
The function estimates sensitivity, specificity, and prevalence parameters, for at least three different methods, without a gold standard.
}
\note{
Please refer to the documentation of the \code{optim} function for method details.
}
\examples{
\dontrun{

ny <- c(1051, 179, 1028, 154, 1040, 159, 981, 208) 
names(ny) <- c('000', '001', '010', '011', '100', '101', '110', '111') 
ny
#  000  001  010  011  100  101  110  111
# 1051  179 1028  154 1040  159  981  208

triangulate(counts = ny, ntest = 3) # no standard error

triangulate(counts = ny, ntest = 3, B = 10) 

}

}
\references{
Yang Z, Xu W, Zhai R, Li T, Ning Z, Pawitan Y, Shen X (2020). Triangulation of analysis strategies links complex traits to specific tissues and cell types. 
\emph{Submitted}.
}
\seealso{
\code{combined.fdr}
}
\author{
Xia Shen, Yudi Pawitan
}
