% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USbb.R
\docType{data}
\name{USbb}
\alias{USbb}
\title{United States of America Domain Boundary Data}
\format{
\code{USbb} is a dataframe with vertices information of the modified external boundaries of the mainland of the United States of America.
\describe{
  \item{x.coord}{The latitudes of the external boundary.}
  \item{y.coord}{The longtitudes of the external boundary.}
}
}
\usage{
data('USbb')
}
\description{
A list of coordinates of the boundary vertices of the United States of America
}
\references{
Data courtesy by First-Data-Lab, 2022.
}
\keyword{datasets}
