% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weird.R
\docType{data}
\name{weird}
\alias{weird}
\title{weird}
\format{\code{weird} is a dataframe with vertices of the boundary of a spiral region.
\describe{
  \item{v1}{The latitudes of the external boundary.}
  \item{v2}{The longtitudes of the external boundary.}
}}
\usage{
data('weird')
}
\description{
A dataframe of coordinates of an external boundary of a spiral region.
}
\references{
This example is courtesy by Phong Luu during class.
}
\keyword{datasets}
